Fix missing semicolon in script_touch-id-bypass.py
Found a missing semicolon in the JavaScript section of the module. Technically the module does work without this semicolon, but if a user were to copy the JavaScript code, remove all new lines (\n) and remove all indents, then run the one liner in Frida, then the user would have issues.

The above scenario is exactly what happened to me.